Refactored code files to simplify directory structure
