--- assets/models/level_test_1.blend
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:63c0068c733a174c2ecabe695230c22121ce8b3b86faddc2d196553074c9fce9
-size 5159781
