+++ assets_src/models/level_test_1.blend
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:e02d45c82afef0f70dc33056dcdf0bb5cfe9286bacf55a76a00358960b034878
-size 5055420
+oid sha256:8ca6841014c15c556042271d290f475fd36ab634b1fb06ffd8a5b0dfe47c9844
+size 5055803
