--- src/assets/mod.rs
@@ -1,7 +1,0 @@
-mod boomerang_assets;
-mod font_assets;
-mod panel_assets;
-
-pub use boomerang_assets::BoomerangAssets;
-pub use font_assets::FontAssets;
-pub use panel_assets::PanelAssets;
