--- src/demo/enemy/enemy_movement.rs
@@ -1,6 +1,0 @@
-use bevy::prelude::ReflectComponent;
-use bevy::prelude::{Component, Reflect};
-
-#[derive(Component, Reflect)]
-#[reflect(Component)]
-pub struct PlayerSeeking;
