--- src/demo/movement.rs
@@ -1,7 +1,0 @@
-use bevy::prelude::*;
-
-#[derive(Component, Debug, Clone, Copy, PartialEq, Default, Reflect)]
-#[reflect(Component)]
-pub struct Velocity {
-    value: Vec3,
-}
