+++ src/screens/gameplay.rs
@@ -2,7 +2,7 @@
 
 use bevy::{input::common_conditions::input_just_pressed, prelude::*};
 
-use crate::{demo::level::spawn_level, screens::Screen};
+use crate::{gameplay::level::spawn_level, screens::Screen};
 
 pub(super) fn plugin(app: &mut App) {
     app.add_systems(OnEnter(Screen::Gameplay), spawn_level);
