+++ src/screens/settings.rs
@@ -4,11 +4,8 @@
 
 use bevy::{audio::Volume, prelude::*, ui::Val::*};
 
-use crate::{
-    assets::{FontAssets, PanelAssets},
-    screens::Screen,
-    theme::prelude::*,
-};
+use crate::ui_assets::{FontAssets, PanelAssets};
+use crate::{screens::Screen, theme::prelude::*};
 
 pub(super) fn plugin(app: &mut App) {
     app.add_systems(OnEnter(Screen::Settings), spawn_settings_screen);
