--- src/settings/mod.rs
@@ -1,7 +1,0 @@
-use bevy::prelude::App;
-
-pub mod framepace;
-
-pub fn plugin(app: &mut App) {
-    app.add_plugins(framepace::plugin);
-}
